+++ packages/better_auth-oauth-provider/test/better_auth/oauth_provider/metadata_test.rb
@@ -19,4 +19,24 @@ class OAuthProviderMetadataTest < Minitest::Test
     assert_equal "http://localhost:3000/api/auth/oauth2/register", body[:registration_endpoint]
     assert_equal ["code"], body[:response_types_supported]
   end
+
+  def test_authorization_server_metadata_omits_registration_endpoint_when_dcr_disabled
+    auth = build_auth(scopes: ["openid"], allow_dynamic_client_registration: false)
+
+    response = auth.api.get_o_auth_server_config(as_response: true)
+    body = JSON.parse(response.body.join, symbolize_names: true)
+
+    assert_equal 200, response.status
+    refute body.key?(:registration_endpoint)
+  end
+
+  def test_openid_metadata_omits_registration_endpoint_when_dcr_disabled
+    auth = build_auth(scopes: ["openid"], allow_dynamic_client_registration: false)
+
+    response = auth.api.get_open_id_config(as_response: true)
+    body = JSON.parse(response.body.join, symbolize_names: true)
+
+    assert_equal 200, response.status
+    refute body.key?(:registration_endpoint)
+  end
 end